backend/requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0

# Supabase
supabase==2.10.0

# Environment variables
python-dotenv==1.0.0

# JWT for authentication
PyJWT==2.8.0

# Password hashing
bcrypt==4.1.2

# Date and time utilities
python-dateutil==2.8.2

# Validation
python-multipart==0.0.6

# Werkzeug for security
Werkzeug==3.0.1
